\name{drS.eg}
\docType{data}
\alias{drS.eg}
\title{A toy similarity matrix for \link{dr}}
\description{
This data set is the toy similarity matrix among the subjects calculated from a toy genotype data set 
which contains of the genotypes of 400 subjects at 200 markers.
}
\usage{data(drS.eg)}
\keyword{datasets}